@@ -16,7 +16,7 @@
 package com.toyproject.booknotes.ui.barcode
 
 import android.content.Context
-import android.support.annotation.UiThread
+import androidx.annotation.UiThread
 import com.google.android.gms.vision.Detector
 import com.google.android.gms.vision.Tracker
 import com.google.android.gms.vision.barcode.Barcode
