@@ -1,7 +1,7 @@
 package com.toyproject.booknotes.extension
 
-import android.arch.lifecycle.Lifecycle
-import android.arch.lifecycle.LifecycleObserver
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
 
 operator fun Lifecycle.plusAssign(observer: LifecycleObserver)
         = this.addObserver(observer)
