@@ -1,16 +1,16 @@
 package com.toyproject.booknotes.rx
 
-import android.arch.lifecycle.Lifecycle
-import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.OnLifecycleEvent
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.OnLifecycleEvent
 import io.reactivex.disposables.Disposable
 
 class AutoActivateDisposable(
         private val lifecycleOwner: LifecycleOwner,
         private val func:()->Disposable
     )
-    :LifecycleObserver {
+    : LifecycleObserver {
 
     private var disposable:Disposable? = null
 
