@@ -1,11 +1,11 @@
 package com.toyproject.booknotes.ui.detail
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.toyproject.booknotes.data.BookInfoDao
 
 class DetailBookInfoViewModelFactory(val bookInfoDao:BookInfoDao)
-    :ViewModelProvider.Factory {
+    : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return DetailBookInfoViewModel(bookInfoDao) as T
     }
