@@ -1,15 +1,16 @@
 package com.toyproject.booknotes.ui.search
 
 import android.content.Context
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+
 import android.util.AttributeSet
 import android.util.Log
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.toyproject.booknotes.api.model.BookInfo
 
 class SearchBookRecyclerView @JvmOverloads
 constructor(context:Context, attrs:AttributeSet?=null, defStyleAtt:Int = 0)
-    :RecyclerView(context, attrs, defStyleAtt){
+    : RecyclerView(context, attrs, defStyleAtt){
 
     internal var isLoading:Boolean = false
 
